requirements.txt
# Telegram Lead Generation Bot - Dependencies
# Python 3.11+ supported. For the smoothest install, use Python 3.11 - 3.13.
# Python 3.14 also works, but requires the newer pinned versions below
# (older pydantic-core / lxml have no 3.14 wheels).

# Telegram bot
python-telegram-bot>=21.6,<22

# Async HTTP
httpx>=0.27,<1.0
aiohttp>=3.10,<4.0

# HTML parsing
# lxml is OPTIONAL at runtime -- we fall back to html.parser if it's missing.
# If you can't build lxml on your platform, comment the next line out.
lxml>=5.3
beautifulsoup4>=4.12,<5.0

# Data handling
pandas>=2.2,<3.0
pydantic>=2.10,<3.0
pydantic-settings>=2.5,<3.0

# Env / config
python-dotenv>=1.0,<2.0

# Retries
tenacity>=9.0,<10.0

# Utility
phonenumbers>=8.13
tldextract>=5.1

# Optional health endpoint
fastapi>=0.115,<1.0
uvicorn>=0.31,<1.0

# Testing
pytest>=8.3,<9.0
pytest-asyncio>=0.24,<1.0
